CORE-8593 Update `GET /analyses` can_share flag with permission check.
The `can_share` flag will now only be set to true if the requesting user
also has `own` permission for the job.
A side effect of this change is that the return values of
`apps.service.apps.job-listings/list-job` and
`apps.service.apps.jobs/update-job-status` will currently always have
this flag set to false, since they do not lookup permissions, but this
`can_share` flag is currently only used in the `GET /analyses` response.