CORE-9016 Handle exceptions in async-submit-batch-jobs.
Updated apps.service.apps.jobs.submissions/async-submit-batch-jobs to
catch and log exceptions, then update the parent job's status to
"Failed" and send a status notification for it.
Also updated this function to examine the status of every sub-job,
and if none successfully submitted, an exception will be thrown which
will also update the parent job's status to "Failed".
A side-effect of these updates means jobs no longer always return the
"Submitted" status, and may now return a "Failed" status.