CORE-1461 Update resource-id-filter fn for abbreviated perm lookups
For CORE-1416, the abbreviated permissions by subject and resource type endpoint was added for performance improvements.  This resulted in the response body from the permissions service changing to a shorter, flatter JSON object.  The resource name is no longer nested in `resource.name`, it's in `resource_name`.  The `resource-id-filter` function just needed to be updated to match this new model.