CORE-1558: fix some SQL syntax errors
This problem was occurring because we're passing `nil` to Korma for some SQL `in`
expressions. This causes Korma to produce a statement that says something like
`WHERE some_value IN NULL`, which isn't valid SQL syntax. We had been using Clojure
expressions such as `[in (seq some-array)]` I believe that the intent was to ensure
that the value passed was a sequence, so the expressions should be changed to
something like `[in (sequence some-array)]`.