CORE-1656 Update PATCH /admin/apps/:system-id/:app-id for app versions
Updated apps.service.apps.de.admin/update-app-details to update the
latest app version, and to mark all versions of an app as deleted or
disabled.

CORE-1655 This also updates the delete behavior for the
`DELETE [/admin]/apps/:system-id/:app-id` endpoints, which will also
now mark all versions of an app as deleted.