CORE-1741 Use client max CPU if lower than tool default
The max CPU request submitted from the client was ignored and not used
in submissions to the job services.
This update will use the max CPU request submitted from the client,
as long as it's less than the tool's default restriction.

Will also use max `memory_limit` from the client in the same way,
if the endpoint is updated in the future to allow it in requests.